reordered commands moved some commands to their own files
